Refactors all endpoint to use same uid variable name.

+++ BudHubServer/del_temp.php
@@ -4,10 +4,42 @@ $user = $credentials["user"];
 $password = $credentials["password"];
 $con = new mysqli("localhost",$user,$password,"budhubdb");
 
-$id = $_GET["id"];
+if ($con->connect_error) {
+    echo $con->connect_error;
+    
+} else {
+    $uid = $_POST["uid"];
+    $tlid = $_POST["lid"];
+    
+    if (!isset($uid) || strlen($uid) < 1 ||
+        !isset($tlid) || strlen($tlid) < 1) {
+        echo "ERROR: Incorrectly formatted data";
+        
+    } else {
+        $st_check = $con->prepare("select * from users where id=?");
+        if (!$st_check) {
+            echo "ERROR: There was a problem verifying the user ID";
 
-$st = $con->prepare("delete from temporder where lid=?");
-$st->bind_param("s",$id);
-$st->execute();
-$rs = $st->get_result();
-echo "1";
+        } else {
+            $st_check->bind_param("s", $uid);
+            $st_check->execute();
+            $rs = $st_check->get_result();
+
+            if($rs->num_rows==0) {
+                echo "ERROR: Invalid user ID";
+
+            } else {
+                $st = $con->prepare("delete from temporder where lid=?");
+                if (!$st_check) {
+                    echo "ERROR: There was a problem querying the database";
+
+                } else {
+                    $st->bind_param("s",$tlid);
+                    $st->execute();
+                    echo "OK";
+                    
+                }
+            }
+        }
+    }
+}
